requirements.txt
Django==1.11.10
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
pip==9.0.1
PyJWT==1.5.3
pytz==2018.3
setuptools==28.8.0

